+++ src/editor/syntax-highlight-renderer.ts
