+++ src/telemetry/document-tracker.ts
